refactor(tests): move archive unit tests to tests/archive integration tests (17/17 pass)
