+++ src-tauri/src/utils/archive.rs
@@ -117,9 +117,9 @@ pub struct TargetResolution {
 
 // ===== 1. 文件头魔数检测 =====
 
-const ZIP_MAGIC: [u8; 4] = [0x50, 0x4B, 0x03, 0x04];
+pub const ZIP_MAGIC: [u8; 4] = [0x50, 0x4B, 0x03, 0x04];
 const ZIP_EMPTY_MAGIC: [u8; 4] = [0x50, 0x4B, 0x05, 0x06];
-const SEVENZ_MAGIC: [u8; 6] = [0x37, 0x7A, 0xBC, 0xAF, 0x27, 0x1C];
+pub const SEVENZ_MAGIC: [u8; 6] = [0x37, 0x7A, 0xBC, 0xAF, 0x27, 0x1C];
 
 pub fn detect_format(path: &Path) -> Result<ArchiveFormat, ArchiveError> {
     let mut f = File::open(path)?;
@@ -678,310 +678,4 @@ fn compress_sevenz(
         ..Default::default()
     });
     Ok(())
-}
-
-// ===== 单测 =====
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn detect_zip() {
-        let tmp = std::env::temp_dir().join("test_detect.zip");
-        std::fs::write(&tmp, ZIP_MAGIC).unwrap();
-        assert_eq!(detect_format(&tmp).unwrap(), ArchiveFormat::Zip);
-        std::fs::remove_file(&tmp).unwrap();
-    }
-
-    #[test]
-    fn detect_7z() {
-        let tmp = std::env::temp_dir().join("test_detect.7z");
-        std::fs::write(&tmp, SEVENZ_MAGIC).unwrap();
-        assert_eq!(detect_format(&tmp).unwrap(), ArchiveFormat::SevenZ);
-        std::fs::remove_file(&tmp).unwrap();
-    }
-
-    #[test]
-    fn detect_unknown() {
-        let tmp = std::env::temp_dir().join("test_detect.bin");
-        std::fs::write(&tmp, b"hello world").unwrap();
-        assert!(detect_format(&tmp).is_err());
-        std::fs::remove_file(&tmp).unwrap();
-    }
-
-    #[test]
-    fn sanitize_blocks_traversal() {
-        assert!(matches!(
-            sanitize_entry_name("../../etc/passwd"),
-            Err(ArchiveError::PathTraversal(_))
-        ));
-        assert!(matches!(
-            sanitize_entry_name("/etc/passwd"),
-            Err(ArchiveError::PathTraversal(_))
-        ));
-        assert!(matches!(
-            sanitize_entry_name("C:\\Windows\\System32"),
-            Err(ArchiveError::PathTraversal(_))
-        ));
-        assert!(matches!(
-            sanitize_entry_name("\\\\server\\share"),
-            Err(ArchiveError::PathTraversal(_))
-        ));
-    }
-
-    #[test]
-    fn sanitize_blocks_macos_metadata() {
-        assert!(matches!(
-            sanitize_entry_name("__MACOSX/foo"),
-            Err(ArchiveError::InvalidName(_))
-        ));
-        assert!(matches!(
-            sanitize_entry_name("role/._background.png"),
-            Err(ArchiveError::InvalidName(_))
-        ));
-        assert!(matches!(
-            sanitize_entry_name("role/.DS_Store"),
-            Err(ArchiveError::InvalidName(_))
-        ));
-    }
-
-    #[test]
-    fn sanitize_clean_path() {
-        assert_eq!(sanitize_entry_name("Alice/bg.png").unwrap(), "Alice/bg.png");
-        assert_eq!(sanitize_entry_name("a:b*c?.txt").unwrap(), "a_b_c_.txt");
-        assert!(sanitize_entry_name("").is_err());
-    }
-
-    #[test]
-    fn safety_allows_large_single_file_without_absolute_limit() {
-        let result = check_entry_safety(0, 2 * 1024 * 1024, 150 * 1024 * 1024);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn safety_allows_large_total_without_absolute_limit() {
-        for index in 0..100 {
-            let result = check_entry_safety(index, 10 * 1024 * 1024, 10 * 1024 * 1024);
-            assert!(result.is_ok(), "entry {index} should pass");
-        }
-    }
-
-    #[test]
-    fn bomb_blocks_high_ratio() {
-        // 100KB 压缩 → 15MB 解压 → 比 153 (> 100)
-        let r = check_entry_safety(0, 100 * 1024, 15 * 1024 * 1024);
-        assert!(matches!(r, Err(ArchiveError::CompressionRatio { .. })));
-    }
-
-    #[test]
-    fn bomb_allows_normal_zip() {
-        for i in 0..100 {
-            let r = check_entry_safety(i, 1024 * 1024, 2 * 1024 * 1024);
-            assert!(r.is_ok(), "entry {i} 应通过");
-        }
-    }
-
-    #[test]
-    fn bomb_blocks_too_many_entries() {
-        let r = check_entry_safety(MAX_ENTRY_COUNT, 0, 0);
-        assert!(matches!(r, Err(ArchiveError::TooManyEntries(_))));
-    }
-
-    #[test]
-    fn resolve_target_skip() {
-        let base = std::env::temp_dir().join(format!("test_resolve_{}", uuid::Uuid::new_v4()));
-        std::fs::create_dir_all(&base).unwrap();
-
-        let r1 = resolve_target(&base, "Alice", ConflictPolicy::Skip).unwrap();
-        assert_eq!(r1.action, "created");
-        std::fs::create_dir_all(&r1.target).unwrap();
-
-        assert!(matches!(
-            resolve_target(&base, "Alice", ConflictPolicy::Skip),
-            Err(ArchiveError::AlreadyExists(_))
-        ));
-
-        let r2 = resolve_target(&base, "Alice", ConflictPolicy::Rename).unwrap();
-        assert_eq!(r2.action, "renamed");
-        assert_eq!(r2.final_name, "Alice_2");
-
-        let r3 = resolve_target(&base, "Alice", ConflictPolicy::Overwrite).unwrap();
-        assert_eq!(r3.action, "overwritten");
-
-        std::fs::remove_dir_all(&base).unwrap();
-    }
-}
-
-
-#[test]
-fn sevenz_round_trip_self_contained() {
-    // 验证 7z 压缩与解压可以完整保留角色顶层目录和必要文件。
-    let tmp = std::env::temp_dir().join(format!("role_7z_{}", uuid::Uuid::new_v4()));
-    let role = _build_sample_role(&tmp);
-    let staging = tmp.join("staging");
-    let wrapped = staging.join("sample");
-    for entry in walkdir(&role) {
-        let rel = entry.strip_prefix(&role).unwrap();
-        let dest = wrapped.join(rel);
-        if entry.is_dir() {
-            std::fs::create_dir_all(&dest).unwrap();
-        } else {
-            std::fs::create_dir_all(dest.parent().unwrap()).unwrap();
-            std::fs::copy(&entry, &dest).unwrap();
-        }
-    }
-    let out_7z = tmp.join("out.7z");
-    sevenz_rust2::compress_to_path(&staging, &out_7z).expect("compress ok");
-    let extract_root = tmp.join("extract");
-    std::fs::create_dir_all(&extract_root).unwrap();
-    let token = CancellationToken::new();
-    let summary = extract_sevenz(&out_7z, &extract_root, &token, &|_| {}).unwrap();
-    assert!(summary.bytes_extracted > 0);
-    assert!(extract_root.join("sample").is_dir());
-    assert!(extract_root.join("sample/settings.yml").exists());
-    std::fs::remove_dir_all(&tmp).unwrap();
-}
-
-fn walkdir(dir: &Path) -> Vec<std::path::PathBuf> {
-    let mut out = Vec::new();
-    if dir.is_file() {
-        out.push(dir.to_path_buf());
-        return out;
-    }
-    for entry in std::fs::read_dir(dir).unwrap() {
-        let entry = entry.unwrap();
-        out.push(entry.path());
-        if entry.path().is_dir() {
-            out.extend(walkdir(&entry.path()));
-        }
-    }
-    out
-}
-
-
-fn _build_sample_role(tmp: &Path) -> PathBuf {
-    let role = tmp.join("sample");
-    std::fs::create_dir_all(role.join("avatar")).unwrap();
-    std::fs::write(role.join("settings.yml"), "title: Sample\n").unwrap();
-    std::fs::write(role.join("avatar/a.png"), b"AAAA").unwrap();
-    std::fs::write(role.join("avatar/b.txt"), b"BBBB").unwrap();
-    role
-}
-
-#[test]
-fn extract_zip_round_trip_self_contained() {
-    // 模拟角色导出目录结构，验证 ZIP 往返后仍保留顶层角色目录。
-    let tmp = std::env::temp_dir().join(format!("role_rt_{}", uuid::Uuid::new_v4()));
-    let role = _build_sample_role(&tmp);
-    let staging = tmp.join("staging");
-    let wrapped = staging.join("sample");
-    for entry in walkdir(&role) {
-        let rel = entry.strip_prefix(&role).unwrap();
-        let dest = wrapped.join(rel);
-        if entry.is_dir() {
-            std::fs::create_dir_all(&dest).unwrap();
-        } else {
-            std::fs::create_dir_all(dest.parent().unwrap()).unwrap();
-            std::fs::copy(&entry, &dest).unwrap();
-        }
-    }
-    // 压缩暂存目录后，压缩包内部应以 `sample/` 作为顶层目录。
-    let out_zip = tmp.join("out.zip");
-    let count = std::sync::atomic::AtomicUsize::new(0);
-    compress(&staging, ArchiveFormat::Zip, &out_zip, &|_| {
-        count.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-    })
-    .unwrap();
-    assert!(out_zip.exists());
-    assert!(count.load(std::sync::atomic::Ordering::SeqCst) > 0);
-
-    let extract_root = tmp.join("extract");
-    std::fs::create_dir_all(&extract_root).unwrap();
-    let token = CancellationToken::new();
-    let summary = extract_zip(&out_zip, &extract_root, &token, &|_| {}).unwrap();
-    assert!(summary.bytes_extracted > 0);
-    assert!(extract_root.join("sample").is_dir());
-    assert!(extract_root.join("sample/settings.yml").exists());
-    std::fs::remove_dir_all(&tmp).unwrap();
-}
-
-#[test]
-fn extract_zip_creates_top_dir_only_when_present() {
-    // 验证带单一顶层角色目录的 ZIP 可以成功生成。
-    let tmp = std::env::temp_dir().join(format!("role_peek_{}", uuid::Uuid::new_v4()));
-    let role = _build_sample_role(&tmp);
-    let out_zip = tmp.join("ok.zip");
-    let staging = tmp.join("staging/sample");
-    for entry in walkdir(&role) {
-        let rel = entry.strip_prefix(&role).unwrap();
-        let dest = staging.join(rel);
-        if entry.is_dir() {
-            std::fs::create_dir_all(&dest).unwrap();
-        } else {
-            std::fs::create_dir_all(dest.parent().unwrap()).unwrap();
-            std::fs::copy(&entry, &dest).unwrap();
-        }
-    }
-    compress(staging.parent().unwrap(), ArchiveFormat::Zip, &out_zip, &|_| {}).unwrap();
-    assert!(out_zip.exists());
-    std::fs::remove_dir_all(&tmp).unwrap();
-}
-
-#[test]
-fn extract_zip_allows_entry_over_legacy_size_limit() {
-    // 已取消绝对大小限制，因此 60MB 条目必须能够完整解压。
-    let tmp = std::env::temp_dir().join(format!("role_bomb_{}", uuid::Uuid::new_v4()));
-    std::fs::create_dir_all(&tmp).unwrap();
-    let staging = tmp.join("staging/sample");
-    std::fs::create_dir_all(&staging).unwrap();
-    use rand::{rngs::StdRng, RngCore, SeedableRng};
-    let mut big = vec![0u8; 60 * 1024 * 1024];
-    StdRng::seed_from_u64(42).fill_bytes(&mut big);
-    std::fs::write(staging.join("big.bin"), &big).unwrap();
-    let out_zip = tmp.join("bomb.zip");
-    compress(staging.parent().unwrap(), ArchiveFormat::Zip, &out_zip, &|_| {}).unwrap();
-    let extract_root = tmp.join("extract");
-    std::fs::create_dir_all(&extract_root).unwrap();
-    let token = CancellationToken::new();
-    let result = extract_zip(&out_zip, &extract_root, &token, &|_| {}).unwrap();
-    assert_eq!(result.files_extracted, 1);
-    assert_eq!(
-        std::fs::metadata(extract_root.join("sample/big.bin"))
-            .unwrap()
-            .len(),
-        60 * 1024 * 1024
-    );
-    std::fs::remove_dir_all(&tmp).unwrap();
-}
-
-#[test]
-fn extract_zip_throttles_progress_events() {
-    // 验证解压过程至少发送开始和完成事件。
-    let tmp = std::env::temp_dir().join(format!("role_throttle_{}", uuid::Uuid::new_v4()));
-    let role = _build_sample_role(&tmp);
-    let staging = tmp.join("staging/sample");
-    for entry in walkdir(&role) {
-        let rel = entry.strip_prefix(&role).unwrap();
-        let dest = staging.join(rel);
-        if entry.is_dir() {
-            std::fs::create_dir_all(&dest).unwrap();
-        } else {
-            std::fs::create_dir_all(dest.parent().unwrap()).unwrap();
-            std::fs::copy(&entry, &dest).unwrap();
-        }
-    }
-    let out_zip = tmp.join("t.zip");
-    compress(staging.parent().unwrap(), ArchiveFormat::Zip, &out_zip, &|_| {}).unwrap();
-
-    let events = std::sync::Mutex::new(Vec::<String>::new());
-    let token = CancellationToken::new();
-    extract_zip(&out_zip, &tmp.join("extract"), &token, &|evt| {
-        events.lock().unwrap().push(evt.phase.to_string());
-    })
-    .unwrap();
-    let phases = events.into_inner().unwrap();
-    assert!(phases.contains(&"started".to_string()));
-    assert!(phases.iter().any(|p| p == "finished"));
-    std::fs::remove_dir_all(&tmp).unwrap();
 }
